+++ CHANGELOG.md
@@ -4,8 +4,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
- * Add methods to set the SSH identity
- * Move connection from the constructor to its own method
+ * Add methods to set the SSH identity. [#11]
+ * Move connection from the constructor to its own method. [#11]
 
 ## [0.1.0] (2021-09-11)
 
@@ -13,3 +13,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 [Unreleased]: https://github.com/ausi/remote-git/compare/0.1.0...HEAD
 [0.1.0]: https://github.com/ausi/remote-git/commits/0.1.0
+
+[#11]: https://github.com/ausi/remote-git/issues/11