+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 
  * Add methods to set the SSH identity
+ * Move connection from the constructor to its own method
 
 ## [0.1.0] (2021-09-11)