+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+
+## [0.2.0] (2021-09-22)
+
  * Add methods to set the SSH identity. [#11]
  * Move connection from the constructor to its own method. [#11]
 
@@ -11,7 +14,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
  * Initial release
 
-[Unreleased]: https://github.com/ausi/remote-git/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/ausi/remote-git/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/ausi/remote-git/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/ausi/remote-git/commits/0.1.0
 
 [#11]: https://github.com/ausi/remote-git/issues/11